+++ src/main/java/de/uros/citlab/errorrate/kws/measures/IRankingMeasure.java
@@ -14,7 +14,7 @@ import java.util.List;
  */
 public interface IRankingMeasure {
 
-    public enum Measure {
+    enum Measure {
         MAP(new MeanAveragePrecision()),
         WMAP(new WMeanAveragePrecision()),
         GAP(new GlobalAveragePrecision()),
@@ -23,11 +23,12 @@ public interface IRankingMeasure {
         R_PRECISION(new RPrecision()),
         G_NCDG(new GNDCG()),
         M_NCDG(new MNDCG()),
-        PRECISION_AT_10(new PrecisionAt10());
+        PRECISION_AT_10(new PrecisionAt10()),
+        AVERAGE_MATCHER_CONFIDENCE(new AverageMatcherConfidence());
 
         private IRankingMeasure method;
 
-        private Measure(IRankingMeasure method) {
+        Measure(IRankingMeasure method) {
             this.method = method;
         }
 
@@ -39,43 +40,4 @@ public interface IRankingMeasure {
 
     public double calcMeasure(List<KWS.MatchList> matchlists);
 
-//    public Stats calcStat(MatchList matches);
-//    public static class Stats {
-//
-//        public int gt_size;
-//        public int falsePositives;
-//        public int falseNegatives;
-//        public int truePositives;
-//        public double measure;
-//        public int queries;
-//
-//        public Stats() {
-//            queries = 0;
-//        }
-//
-//        Stats(double measure, int truePositives, int falsePositives, int falseNegatives, int gt_size) {
-//            this.measure = measure;
-//            this.truePositives = truePositives;
-//            this.falsePositives = falsePositives;
-//            this.falseNegatives = falseNegatives;
-//            this.gt_size = gt_size;
-//        }
-//
-//        public void accumulate(Stats calcStat) {
-//            queries++;
-//            falseNegatives += calcStat.falseNegatives;
-//            falsePositives += calcStat.falsePositives;
-//            measure += calcStat.gt_size == 0
-//                    ? (falsePositives == 0 ? 1 : 0)
-//                    : (double) calcStat.truePositives / calcStat.gt_size;
-//            gt_size += calcStat.gt_size;
-//            truePositives += calcStat.truePositives;
-//        }
-//
-//        @Override
-//        public String toString() {
-//            return "Stats{" + "gt_size=" + gt_size + ", falsePositives=" + falsePositives + ", falseNegatives=" + falseNegatives + ", truePositives=" + truePositives + '}';
-//        }
-//
-//    }
 }